+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 
 ## master
 
-* Fixed interface with GitLab CI [@pbendersky](https://github.com/pbendersky)
+## 5.12.0
 
+* Fixed interface with GitLab CI [@pbendersky](https://github.com/pbendersky)
 * `danger pr` now supports Bitbucket Cloud repos [@RicardoBelchior](https://github.com/RicardoBelchior)
 
 ## 5.11.1